linkedin_scraper/requirements.txt
# Web scraping and browser automation
selenium>=4.15.0
playwright>=1.40.0
beautifulsoup4>=4.12.0

# Data processing and export
pandas>=2.1.0

# Configuration and environment
python-dotenv>=1.0.0

# Web interface
flask>=3.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Error tracking (optional)
# Uncomment to enable Sentry integration
# sentry-sdk>=1.40.0
